TAN-89 / CT-18: split lib.rs to satisfy file-size gate
lib.rs was 3502 lines; touching it (to register the authority module)
tripped the 2000-line hard cap. Extract the connector/source/ingestion
type block and the test module into include!()'d part files
(lib_sources.rs, lib_tests.rs), preserving the flat public API. lib.rs is
now 1698 lines; all 48 contract-crate tests still pass.
